backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
lightrag==0.0.0a13
openai==1.3.5
python-dotenv>=1.0.1
aiofiles==23.2.1
python-multipart==0.0.6
